Move review reaction Lottie files into folder
