+++ apps/web/src/screens/review/reactions/lottie/assets/review_again_snail.json
