+++ apps/web/src/screens/review/reactions/lottie/assets/review_easy_plant.json
