+++ apps/web/src/screens/review/reactions/lottie/assets/review_good_seal.json
