+++ apps/web/src/screens/review/reactions/lottie/assets/review_hard_rooster.json
